builds on gcc again
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@1519 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -13623,7 +13623,7 @@ bool CSphIndex_VLN::SetupQueryWord ( CSphQueryWord & tWord, const CSphTermSetup 
 		if ( sphSeek ( tTermSetup.m_tWordlist.GetFD (), uWordlistOffset, SEEK_SET ) < 0 )
 			return false;
 
-		if ( sphReadThrottled ( tTermSetup.m_tWordlist.GetFD (), m_pWordlistChunkBuf, (size_t)iChunkLength ) != iChunkLength )
+		if ( sphReadThrottled ( tTermSetup.m_tWordlist.GetFD (), m_pWordlistChunkBuf, (size_t)iChunkLength ) != (size_t)iChunkLength )
 			return false;
 
 		pBuf = m_pWordlistChunkBuf;
